chore: #94 move block utils to separate file
